+++ src/components/landelijk/intake-hospital-barscale.tsx
@@ -1,0 +1,43 @@
+import BarScale from '@/components/barScale';
+
+import { IntakeHospitalMa } from '@/types/data.d';
+import siteText from '@/locale/index';
+
+const text: typeof siteText.ziekenhuisopnames_per_dag =
+  siteText.ziekenhuisopnames_per_dag;
+
+export function IntakeHospitalBarScale(props: {
+  data: IntakeHospitalMa | undefined;
+  showAxis: boolean;
+}) {
+  const { data, showAxis } = props;
+
+  if (!data) return null;
+
+  return (
+    <BarScale
+      min={0}
+      max={100}
+      signaalwaarde={40}
+      screenReaderText={text.barscale_screenreader_text}
+      value={data.last_value.moving_average_hospital}
+      id="opnames"
+      rangeKey="moving_average_hospital"
+      gradient={[
+        {
+          color: '#69c253',
+          value: 0,
+        },
+        {
+          color: '#D3A500',
+          value: 40,
+        },
+        {
+          color: '#f35065',
+          value: 90,
+        },
+      ]}
+      showAxis={showAxis}
+    />
+  );
+}
