+++ src/components/layout/NationalLayout.tsx
@@ -4,16 +4,18 @@ import { useRouter } from 'next/router';
 
 import TitleWithIcon from '@/components/titleWithIcon';
 import { getLayout as getSiteLayout } from '@/components/layout';
-import { ReproductionIndexBarScale } from '@/pages/landelijk/reproductiegetal';
-import { PostivelyTestedPeopleBarScale } from '@/pages/landelijk/positief-geteste-mensen';
-import { InfectiousPeopleBarScale } from '@/pages/landelijk/besmettelijke-mensen';
-import { IntakeHospitalBarScale } from '@/pages/landelijk/ziekenhuis-opnames';
-import { IntakeIntensiveCareBarscale } from '@/pages/landelijk/intensive-care-opnames';
-import { SuspectedPatientsBarScale } from '@/pages/landelijk/verdenkingen-huisartsen';
-import { SewerWaterBarScale } from '@/pages/landelijk/rioolwater';
-import { NursingHomeInfectedPeopleBarScale } from '@/pages/landelijk/verpleeghuis-positief-geteste-personen';
-import { NursingHomeInfectedLocationsBarScale } from '@/pages/landelijk/verpleeghuis-besmette-locaties';
-import { NursingHomeDeathsBarScale } from '@/pages/landelijk/verpleeghuis-sterfte';
+
+import { ReproductionIndexBarScale } from '@/components/landelijk/reproduction-index-barscale';
+import { PositiveTestedPeopleBarScale } from '@/components/landelijk/positive-tested-people-barscale';
+import { InfectiousPeopleBarScale } from '@/components/landelijk/infectious-people-barscale';
+import { IntakeHospitalBarScale } from '@/components/landelijk/intake-hospital-barscale';
+
+import { IntakeIntensiveCareBarscale } from '@/components/landelijk/intake-intensive-care-barscale';
+import { SuspectedPatientsBarScale } from '@/components/landelijk/suspected-patients-barscale';
+import { SewerWaterBarScale } from '@/components/landelijk/sewer-water-barscale';
+import { NursingHomeInfectedPeopleBarScale } from '@/components/landelijk/nursing-home-infected-people-barscale';
+import { NursingHomeInfectedLocationsBarScale } from '@/components/landelijk/nursing-home-infected-locations-barscale';
+import { NursingHomeDeathsBarScale } from '@/components/landelijk/nursing-home-deaths-barscale';
 
 import GetestIcon from '@/assets/test.svg';
 import ReproIcon from '@/assets/reproductiegetal.svg';
@@ -139,7 +141,7 @@ function NationalLayout(props: WithChildren<INationalData>) {
                       title={siteText.positief_geteste_personen.titel_sidebar}
                     />
                     <span>
-                      <PostivelyTestedPeopleBarScale
+                      <PositiveTestedPeopleBarScale
                         data={data?.infected_people_delta_normalized}
                         showAxis={true}
                       />
