+++ src/components/screenReaderOnly/index.tsx
@@ -1,5 +1,5 @@
 import styles from './screenReaderOnly.module.scss';
-import { WithChildren } from '@types/index';
+import { WithChildren } from '@/types/index';
 
 export default ScreenReaderOnly;
