+++ src/pages/landelijk/intensive-care-opnames.tsx
@@ -1,8 +1,8 @@
-import BarScale from '@/components/barScale';
 import { FCWithLayout } from '@/components/layout';
 import { getNationalLayout } from '@/components/layout/NationalLayout';
 import { LineChart } from '@/components/charts/index';
 import { ContentHeader } from '@/components/layout/Content';
+import { IntakeIntensiveCareBarscale } from '@/components/landelijk/intake-intensive-care-barscale';
 
 import Arts from '@/assets/arts.svg';
 
@@ -13,42 +13,6 @@ import { IntakeIntensivecareMa } from '@/types/data.d';
 import getNlData, { INationalData } from '@/static-props/nl-data';
 
 const text: typeof siteText.ic_opnames_per_dag = siteText.ic_opnames_per_dag;
-
-export function IntakeIntensiveCareBarscale(props: {
-  data: IntakeIntensivecareMa | undefined;
-  showAxis: boolean;
-}) {
-  const { data, showAxis } = props;
-
-  if (!data) return null;
-
-  return (
-    <BarScale
-      min={0}
-      max={30}
-      gradient={[
-        {
-          color: '#69c253',
-          value: 0,
-        },
-        {
-          color: '#D3A500',
-          value: 10,
-        },
-        {
-          color: '#f35065',
-          value: 20,
-        },
-      ]}
-      rangeKey="moving_average_ic"
-      screenReaderText={text.barscale_screenreader_text}
-      signaalwaarde={10}
-      value={data.last_value.moving_average_ic}
-      id="ic"
-      showAxis={showAxis}
-    />
-  );
-}
 
 const IntakeIntensiveCare: FCWithLayout<INationalData> = (props) => {
   const { data: state } = props;
