+++ src/pages/landelijk/reproductiegetal.tsx
@@ -1,9 +1,9 @@
-import BarScale from '@/components/barScale';
 import Legenda from '@/components/legenda';
 import { FCWithLayout } from '@/components/layout';
 import { getNationalLayout } from '@/components/layout/NationalLayout';
 import { AreaChart } from '@/components/charts/index';
 import { ContentHeader } from '@/components/layout/Content';
+import { ReproductionIndexBarScale } from '@/components/landelijk/reproduction-index-barscale';
 
 import Repro from '@/assets/reproductiegetal.svg';
 
@@ -14,47 +14,6 @@ import { ReproductionIndex as ReproductionIndexData } from '@/types/data.d';
 import getNlData, { INationalData } from '@/static-props/nl-data';
 
 const text: typeof siteText.reproductiegetal = siteText.reproductiegetal;
-
-export function ReproductionIndexBarScale(props: {
-  data: ReproductionIndexData | undefined;
-  lastKnown: ReproductionIndexData | undefined;
-  showAxis: boolean;
-}) {
-  const { data, lastKnown, showAxis } = props;
-
-  if (!data) return null;
-
-  return (
-    <BarScale
-      min={0}
-      max={2}
-      screenReaderText={text.barscale_screenreader_text}
-      signaalwaarde={1}
-      value={lastKnown?.last_value?.reproduction_index_avg}
-      id="repro"
-      rangeKey="reproduction_index_avg"
-      gradient={[
-        {
-          color: '#69c253',
-          value: 0,
-        },
-        {
-          color: '#69c253',
-          value: 1,
-        },
-        {
-          color: '#D3A500',
-          value: 1.0104,
-        },
-        {
-          color: '#f35065',
-          value: 1.125,
-        },
-      ]}
-      showAxis={showAxis}
-    />
-  );
-}
 
 const ReproductionIndex: FCWithLayout<INationalData> = (props) => {
   const { data: state } = props;
