+++ src/pages/landelijk/rioolwater.tsx
@@ -1,8 +1,9 @@
-import BarScale from '@/components/barScale';
 import { FCWithLayout } from '@/components/layout';
 import { getNationalLayout } from '@/components/layout/NationalLayout';
 import { LineChart } from '@/components/charts/index';
 import { ContentHeader } from '@/components/layout/Content';
+
+import { SewerWaterBarScale } from '@/components/landelijk/sewer-water-barscale';
 
 import RioolwaterMonitoring from '@/assets/rioolwater-monitoring.svg';
 
@@ -13,33 +14,6 @@ import { RioolwaterMetingen } from '@/types/data.d';
 import getNlData, { INationalData } from '@/static-props/nl-data';
 
 const text: typeof siteText.rioolwater_metingen = siteText.rioolwater_metingen;
-
-export function SewerWaterBarScale(props: {
-  data: RioolwaterMetingen | undefined;
-  showAxis: boolean;
-}) {
-  const { data, showAxis } = props;
-
-  if (!data) return null;
-
-  return (
-    <BarScale
-      min={0}
-      max={100}
-      screenReaderText={text.barscale_screenreader_text}
-      value={Number(data.last_value.average)}
-      id="rioolwater_metingen"
-      rangeKey="average"
-      gradient={[
-        {
-          color: '#3391CC',
-          value: 0,
-        },
-      ]}
-      showAxis={showAxis}
-    />
-  );
-}
 
 const SewerWater: FCWithLayout<INationalData> = (props) => {
   const { data: state } = props;
