+++ src/pages/landelijk/verdenkingen-huisartsen.tsx
@@ -1,8 +1,9 @@
-import BarScale from '@/components/barScale';
 import { ContentHeader } from '@/components/layout/Content';
 import { FCWithLayout } from '@/components/layout';
 import { getNationalLayout } from '@/components/layout/NationalLayout';
 import { LineChart } from '@/components/charts/index';
+
+import { SuspectedPatientsBarScale } from '@/components/landelijk/suspected-patients-barscale';
 
 import Arts from '@/assets/arts.svg';
 
@@ -15,33 +16,6 @@ import getNlData, { INationalData } from '@/static-props/nl-data';
 
 const text: typeof siteText.verdenkingen_huisartsen =
   siteText.verdenkingen_huisartsen;
-
-export function SuspectedPatientsBarScale(props: {
-  data: VerdenkingenHuisartsen | undefined;
-  showAxis: boolean;
-}) {
-  const { data, showAxis } = props;
-
-  if (!data) return null;
-
-  return (
-    <BarScale
-      min={0}
-      max={140}
-      screenReaderText={text.barscale_screenreader_text}
-      value={data.last_value.incidentie as number | null}
-      id="verdenkingen_huisartsen"
-      rangeKey="incidentie"
-      gradient={[
-        {
-          color: '#3391CC',
-          value: 0,
-        },
-      ]}
-      showAxis={showAxis}
-    />
-  );
-}
 
 const SuspectedPatients: FCWithLayout<INationalData> = (props) => {
   const { data: state } = props;
