+++ src/pages/landelijk/verpleeghuis-sterfte.tsx
@@ -1,8 +1,9 @@
-import BarScale from '@/components/barScale';
 import { ContentHeader } from '@/components/layout/Content';
 import { FCWithLayout } from '@/components/layout';
 import { getNationalLayout } from '@/components/layout/NationalLayout';
 import { LineChart } from '@/components/charts/index';
+
+import { NursingHomeDeathsBarScale } from '@/components/landelijk/nursing-home-deaths-barscale';
 
 import CoronaVirus from '@/assets/coronavirus.svg';
 
@@ -13,33 +14,6 @@ import getNlData, { INationalData } from '@/static-props/nl-data';
 
 const text: typeof siteText.verpleeghuis_oversterfte =
   siteText.verpleeghuis_oversterfte;
-
-export function NursingHomeDeathsBarScale(props: {
-  data: DeceasedPeopleNurseryCountDaily | undefined;
-  showAxis: boolean;
-}) {
-  const { data, showAxis } = props;
-
-  if (!data) return null;
-
-  return (
-    <BarScale
-      min={0}
-      max={50}
-      screenReaderText={text.barscale_screenreader_text}
-      value={data.last_value.deceased_nursery_daily}
-      id="over"
-      rangeKey="deceased_nursery_daily"
-      gradient={[
-        {
-          color: '#3391CC',
-          value: 0,
-        },
-      ]}
-      showAxis={showAxis}
-    />
-  );
-}
 
 const NursingHomeDeaths: FCWithLayout<INationalData> = (props) => {
   const { data: state } = props;
