+++ src/pages/landelijk/ziekenhuis-opnames.tsx
@@ -1,8 +1,9 @@
-import BarScale from '@/components/barScale';
 import { FCWithLayout } from '@/components/layout';
 import { getNationalLayout } from '@/components/layout/NationalLayout';
 import { LineChart } from '@/components/charts/index';
 import { ContentHeader } from '@/components/layout/Content';
+
+import { IntakeHospitalBarScale } from '@/components/landelijk/intake-hospital-barscale';
 
 import Ziekenhuis from '@/assets/ziekenhuis.svg';
 
@@ -44,42 +45,6 @@ const tooltipRegionContent = (context: any): ReactNode => {
     )
   );
 };
-
-export function IntakeHospitalBarScale(props: {
-  data: IntakeHospitalMa | undefined;
-  showAxis: boolean;
-}) {
-  const { data, showAxis } = props;
-
-  if (!data) return null;
-
-  return (
-    <BarScale
-      min={0}
-      max={100}
-      signaalwaarde={40}
-      screenReaderText={text.barscale_screenreader_text}
-      value={data.last_value.moving_average_hospital}
-      id="opnames"
-      rangeKey="moving_average_hospital"
-      gradient={[
-        {
-          color: '#69c253',
-          value: 0,
-        },
-        {
-          color: '#D3A500',
-          value: 40,
-        },
-        {
-          color: '#f35065',
-          value: 90,
-        },
-      ]}
-      showAxis={showAxis}
-    />
-  );
-}
 
 const IntakeHospital: FCWithLayout<INationalData> = (props) => {
   const { data: state } = props;
